+++ src/integrations/lottie/plugin.rs
@@ -17,14 +17,11 @@ impl Plugin for LottieIntegrationPlugin {
     fn build(&self, app: &mut App) {
         app.init_asset_loader::<VelloLottieLoader>()
             .init_asset::<VelloLottie>()
+            .add_systems(PostUpdate, systems::advance_playheads)
             .add_systems(
-                PostUpdate,
-                (
-                    systems::advance_playheads_without_options,
-                    systems::advance_playheads_with_options,
-                ),
+                Last,
+                (systems::run_transitions, systems::transition_state).chain(),
             )
-            .add_systems(Last, systems::spawn_playheads)
             .add_systems(
                 PostUpdate,
                 check_visibility::<With<VelloLottieHandle>>